# docs/openapi/swagger.yaml
components:
  schemas:
...
    ? github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Data-github_com_Mininglamp-OSS_octo-marketplace_internal_service_skill_SkillItem
    : properties:
        data:
          $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_service_skill.SkillItem'
      type: object
...
    github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Data-internal_api_handler_plugin_delistResponse:
      properties:
        data:
          $ref: '#/components/schemas/internal_api_handler_plugin.delistResponse'
      type: object
...
    github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Data-internal_api_handler_plugin_publishResponse:
      properties:
        data:
          $ref: '#/components/schemas/internal_api_handler_plugin.publishResponse'
      type: object
...
    github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error:
      properties:
        error:
          $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.ErrorBody'
      type: object
    github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.ErrorBody:
      properties:
        code:
          type: string
        details:
          additionalProperties: {}
          type: object
        hint:
          type: string
        message:
          type: string
      type: object
...
    github_com_Mininglamp-OSS_octo-marketplace_internal_model.PluginDisplayStatus:
      description: 'DisplayStatus is the single DERIVED status a client shows: draft
        /

        pending_review / published / rejected / delisted. It folds the listing axis

        together with the review axis so clients do not have to fetch the review list

        and join. Distinct from `status`, which is the row''s soft-active flag.'
      enum:
      - draft
      - pending_review
      - published
      - rejected
      - delisted
      type: string
      x-enum-comments:
        PluginDisplayStatusDelisted: 已下架
        PluginDisplayStatusDraft: 草稿
        PluginDisplayStatusPendingReview: 审核中
        PluginDisplayStatusPublished: 已发布
        PluginDisplayStatusRejected: 驳回
      x-enum-varnames:
      - PluginDisplayStatusDraft
      - PluginDisplayStatusPendingReview
      - PluginDisplayStatusPublished
      - PluginDisplayStatusRejected
      - PluginDisplayStatusDelisted
    github_com_Mininglamp-OSS_octo-marketplace_internal_model.PluginListingState:
      description: 'PluginListingState is the plugin''s CURRENT listing state, which
        can have

        moved on since this request was decided — an approved plugin a Space admin

        later delisted reads `delisted`. The review queue renders the row status from

        it and uses it to stop offering 下架 on something already down.'
      enum:
      - draft
      - published
      - delisted
      type: string
      x-enum-varnames:
      - PluginListingStateDraft
      - PluginListingStatePublished
      - PluginListingStateDelisted
...
    github_com_Mininglamp-OSS_octo-marketplace_internal_service_skill.SkillItem:
      properties:
        category_id:
          type: string
        created_at:
          type: string
        creator_id:
          type: string
        creator_name:
          type: string
        description:
          type: string
        display_name:
          type: string
        download_count:
          type: integer
        file_name:
          type: string
        file_size:
          type: integer
        icon_url:
          type: string
        name:
          type: string
        owner_name:
          type: string
        readme_content:
          type: string
        skill_id:
          type: string
        tags:
          items:
            type: string
          type: array
          uniqueItems: false
        updated_at:
          type: string
        version:
          type: string
        view_count:
          type: integer
        visibility:
          type: string
      type: object
...
    internal_api_handler_plugin.delistRequest:
      properties:
        plugin_id:
          type: string
        reason:
          type: string
      type: object
    internal_api_handler_plugin.delistResponse:
      properties:
        display_status:
          $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_model.PluginDisplayStatus'
        listing_state:
          $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_model.PluginListingState'
        plugin_id:
          type: string
      type: object
...
    internal_api_handler_plugin.publishRequest:
      properties:
        changelog:
          type: string
        plugin_id:
          type: string
        version:
          description: 'Version and Changelog are used only when the publish needs
            organization

            review; both are ignored for a 仅自己可见 Plugin, which lists immediately.

            Version defaults to the draft''s current version.'
          type: string
      type: object
    internal_api_handler_plugin.publishResponse:
      properties:
        display_status:
          $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_model.PluginDisplayStatus'
        listing_state:
          $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_model.PluginListingState'
        plugin_id:
          type: string
        review_id:
          description: 'ReviewID is present only when the publish opened a review
            request, which is

            how the client knows which branch fired without a second call.'
          type: string
      type: object
...
    internal_api_handler_upload.BotPublishSkillRequest:
      properties:
        category_id:
          type: string
        changelog:
          type: string
        description:
          type: string
        display_name:
          type: string
        icon_url:
          type: string
        name:
          type: string
        presigned_url:
          type: string
        publish_mode:
          type: string
        skill_upload_id:
          type: string
        tags:
          items:
            maxLength: 10
            type: string
          maxItems: 10
          type: array
          uniqueItems: false
        upload_url:
          type: string
        version:
          type: string
        visibility:
          type: string
      type: object
...
  securitySchemes:
    AdminToken:
      in: header
      name: X-Admin-Token
      type: apiKey
    Bearer:
      scheme: bearer
      type: http
info:
  contact:
    name: OCTO API Team
    url: https://github.com/Mininglamp-OSS/octo-marketplace
  description: 'Skill and MCP marketplace API for OCTO.


    A 409 CONFLICT carries `error.details.conflict_reason`, which is what a

    client should branch on — the status code alone does not distinguish these:

    `review_pending` (a request is already open on this plugin; cancel it first),

    `listed_requires_review` (the plugin is listed to the organization, so the

    change must go through review or an admin must delist it first),

    `label_taken` (the submitted version label is already published; pick another

    and resubmit), `already_published` / `not_published` (the listing transition

    does not apply in the plugin''s current state), `deadlock` (transient lock

    contention; also sets `retryable: true`, so retry the request unchanged), and

    `state` (a generic raced state change; refresh and retry).'
  title: Octo Marketplace API
  version: 1.0.0
openapi: 3.1.0
paths:
...
  /bot/skills/publish:
    post:
      description: Parse an uploaded Skill archive synchronously through the bounded
        worker pool, then create a Skill owned by the bot owner.
      operationId: bot_skill.publish
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/internal_api_handler_upload.BotPublishSkillRequest'
                description: Bot publish request
                summary: body
        description: Bot publish request
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Data-github_com_Mininglamp-OSS_octo-marketplace_internal_service_skill_SkillItem'
          description: Created
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: VALIDATION_ERROR
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: AUTH_REQUIRED
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: FORBIDDEN
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: NOT_FOUND
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: CONFLICT
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: INTERNAL_ERROR
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: UPSTREAM_UNAVAILABLE
      security:
      - Bearer: []
      summary: Publish Skill as bot
      tags:
      - skill_upload
...
  /plugins/delist:
    post:
      description: 'Take a published Plugin out of the marketplace. Requires the Space
        owner or admin role — the same authority as approving a review request. Only
        an organization-visible, standalone Plugin can be delisted: a private Plugin
        is not organization content and an embedded child is listed and un-listed
        by its container, so both answer 404 rather than confirming they exist. Any
        pending review request on the Plugin is canceled. The Plugin stays editable
        and can be published again by its owner.'
      operationId: plugin.delist
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/internal_api_handler_plugin.delistRequest'
                description: Plugin ID and an optional reason
                summary: body
        description: Plugin ID and an optional reason
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Data-internal_api_handler_plugin_delistResponse'
          description: OK
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: VALIDATION_ERROR
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: AUTH_REQUIRED
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: FORBIDDEN
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: NOT_FOUND
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: CONFLICT
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: INTERNAL_ERROR
      security:
      - Bearer: []
      summary: Delist plugin
      tags:
      - plugin
...
  /plugins/publish:
    post:
      description: 'Publish a caller-owned Plugin. The Plugin''s visibility decides
        what that means: a private Plugin is listed immediately, while an organization-visible
        Plugin opens a review request and stays a draft until a Space admin approves
        it. There is no separate submit-for-review endpoint.'
      operationId: plugin.publish
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/internal_api_handler_plugin.publishRequest'
                description: Plugin ID, with an optional version label and changelog
                  for the review branch
                summary: body
        description: Plugin ID, with an optional version label and changelog for the
          review branch
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Data-internal_api_handler_plugin_publishResponse'
          description: OK
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: VALIDATION_ERROR
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: AUTH_REQUIRED
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: NOT_FOUND
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: CONFLICT
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_Mininglamp-OSS_octo-marketplace_internal_api_response.Error'
          description: INTERNAL_ERROR
      security:
      - Bearer: []
      summary: Publish plugin
      tags:
      - plugin
...
servers:
- url: /v1
...